feat: implement modular DOCX generators and collection management for document export
