+++ packages/docx_creator/lib/src/exporters/docx/docx_collection_manager.dart
@@ -1,0 +1,168 @@
+import 'package:docx_creator/docx_creator.dart';
+import 'docx_export_state.dart';
+
+/// Pre-processes the document before generation to collect and catalogue items
+/// like images, bullet lists, fonts, and sets up appropriate counters/IDs.
+class DocxCollectionManager {
+  static void collect(DocxExportState state) {
+    // Register document fonts
+    for (var font in state.doc.fonts) {
+      state.fontManager.registerFont(font);
+    }
+
+    // Process background image
+    if (state.doc.section?.backgroundImage != null) {
+      state.backgroundImage = state.doc.section!.backgroundImage;
+      state.backgroundImage!.setRelationshipId('rIdBg');
+      final ext = state.backgroundImage!.normalizedExtension;
+      state.images['word/media/background.$ext'] = state.backgroundImage!.bytes;
+    }
+
+    // Process images recursively
+    _collectImagesGrouped(state);
+
+    final allImages = <DocxInlineImage>{
+      ...state.groupedImages['body']!,
+      ...state.groupedImages['header']!,
+      ...state.groupedImages['footer']!,
+    };
+
+    for (var img in allImages) {
+      state.imageCounter++;
+      final rId = 'rId${state.imageCounter + 10}';
+      img.setRelationshipId(rId, state.uniqueIdCounter++);
+      final mediaPath =
+          'word/media/image${state.imageCounter}.${img.extension}';
+      state.imageMediaPaths[img] = mediaPath;
+      state.images[mediaPath] = img.bytes;
+    }
+
+    // Process lists recursively
+    _collectLists(state);
+  }
+
+  static void _collectImagesGrouped(DocxExportState state) {
+    final bodyImages = <DocxInlineImage>[];
+    final headerImages = <DocxInlineImage>[];
+    final footerImages = <DocxInlineImage>[];
+
+    for (var element in state.doc.elements) {
+      _collectImagesFromNode(element, bodyImages);
+    }
+    if (state.doc.section?.header != null) {
+      for (var child in state.doc.section!.header!.children) {
+        _collectImagesFromNode(child, headerImages);
+      }
+    }
+    if (state.doc.section?.footer != null) {
+      for (var child in state.doc.section!.footer!.children) {
+        _collectImagesFromNode(child, footerImages);
+      }
+    }
+
+    state.groupedImages = {
+      'body': bodyImages,
+      'header': headerImages,
+      'footer': footerImages,
+    };
+  }
+
+  static void _collectImagesFromNode(
+      DocxNode node, List<DocxInlineImage> images) {
+    if (node is DocxImage) {
+      images.add(node.asInline);
+    } else if (node is DocxInlineImage) {
+      images.add(node);
+    } else if (node is DocxParagraph) {
+      for (var child in node.children) {
+        _collectImagesFromNode(child, images);
+      }
+    } else if (node is DocxTable) {
+      for (var row in node.rows) {
+        for (var cell in row.cells) {
+          for (var child in cell.children) {
+            _collectImagesFromNode(child, images);
+          }
+        }
+      }
+    } else if (node is DocxHeader) {
+      for (var child in node.children) {
+        _collectImagesFromNode(child, images);
+      }
+    } else if (node is DocxFooter) {
+      for (var child in node.children) {
+        _collectImagesFromNode(child, images);
+      }
+    }
+  }
+
+  static void _collectLists(DocxExportState state) {
+    final allLists = <DocxList>[];
+    for (var element in state.doc.elements) {
+      _collectListsFromNode(element, allLists);
+    }
+
+    int abstractNumIdCounter = 2; // 0 and 1 are reserved for default styles
+
+    for (var list in allLists) {
+      int exportedNumId;
+      final sourceNumId = list.numId;
+
+      if (sourceNumId != null &&
+          state.preservedNumIds.containsKey(sourceNumId)) {
+        // Reuse existing exported ID for continuity
+        exportedNumId = state.preservedNumIds[sourceNumId]!;
+      } else {
+        // Create new exported ID
+        exportedNumId = state.numIdCounter++;
+        if (sourceNumId != null) {
+          state.preservedNumIds[sourceNumId] = exportedNumId;
+        }
+
+        if (list.style.imageBulletBytes != null) {
+          // Image Bullet List
+          final bulletIndex = state.imageBullets.length;
+          state.imageBullets.add(list.style.imageBulletBytes!);
+
+          final absId = abstractNumIdCounter++;
+          state.abstractNumImageBulletMap[absId] = bulletIndex;
+          state.listAbstractNumMap[exportedNumId] = absId;
+        } else {
+          // Standard List
+          state.listAbstractNumMap[exportedNumId] = list.isOrdered ? 1 : 0;
+          // Only apply start override if this is the start of the chain (new ID)
+          if (list.isOrdered && list.startIndex > 1) {
+            state.listStartOverrides[exportedNumId] = list.startIndex;
+          }
+        }
+      }
+
+      list.numId = exportedNumId;
+    }
+  }
+
+  static void _collectListsFromNode(DocxNode node, List<DocxList> lists) {
+    if (node is DocxList) {
+      lists.add(node);
+      // Also collect nested lists within list items
+      for (var item in node.items) {
+        for (var child in item.children) {
+          _collectListsFromNode(child, lists);
+        }
+      }
+    } else if (node is DocxTable) {
+      for (var row in node.rows) {
+        for (var cell in row.cells) {
+          for (var child in cell.children) {
+            _collectListsFromNode(child, lists);
+          }
+        }
+      }
+    } else if (node is DocxParagraph) {
+      // Paragraphs might contain inline elements with nested content
+      for (var child in node.children) {
+        _collectListsFromNode(child, lists);
+      }
+    }
+  }
+}
